Novas RFs

+++ database/migrations/2026_06_02_205837_create_items_table.php
@@ -19,8 +19,8 @@ return new class extends Migration
             $table->boolean('is_for_sale')->default(false); // Flag para exibir no catálogo público [cite: 25]
             $table->decimal('cost_price', 10, 2)->default(0);
             $table->decimal('sale_price', 10, 2)->default(0);
-            $table->integer('stock_quantity')->default(0);
-            $table->integer('min_stock_alert')->default(5); // Para o Dashboard (RF-10) 
+            $table->decimal('stock_quantity', 10, 2)->default(0);
+            $table->decimal('min_stock_alert', 10, 2)->default(5); // Para o Dashboard (RF-10) 
             $table->softDeletes();
             $table->timestamps();
